backend/requirements.txt
# Core API and Web Server
fastapi==0.129.0
uvicorn[standard]==0.40.0
gunicorn==25.0.2
python-multipart==0.0.22

# Data Processing and Science
pandas==3.0.0
numpy==2.4.2
scikit-learn==1.8.0
scipy==1.17.0
statsmodels==0.14.6
openpyxl==3.1.5

# AI and Integrations
google-genai>=1.0.0
openai>=1.0.0
tenacity==9.1.4

# Database and Utils
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
python-dotenv==1.2.1
email-validator==2.3.0
supabase>=2.15.0
PyJWT==2.10.0
python-jose==3.3.0

# Frontend Assets (if served by backend)
watchfiles==1.1.1

# Flask (for legacy or alternative entry points)
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.5
